pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["jasonjiangstuy <jasonjiangny@gmail.com>"]
description = ""
name = "stuyescalator"
version = "0.1.0"
[tool.poetry.dependencies]
flask = "^1.1"
python = "^3.7"
google-cloud = "^0.34.0"
google-cloud-datastore = "^1.11"
google-nucleus = "^0.5.0"
[tool.poetry.dev-dependencies]
